+++ src/form-accessor-base.ts
@@ -1,7 +1,6 @@
 import { computed } from "mobx";
 import { FormDefinition, GroupDefinition } from "./form";
 import {
-  Accessor,
   FieldAccess,
   RepeatingFormAccess,
   SubFormAccess,
@@ -9,6 +8,8 @@ import {
 } from "./accessor";
 import { FormAccessor } from "./form-accessor";
 import { ValidateOptions } from "./validate-options";
+import { ExternalMessages } from "./validationMessages";
+import { IAccessor } from "./interfaces";
 
 // a base class that delegates to a form accessor
 export abstract class FormAccessorBase<
@@ -55,11 +56,50 @@ export abstract class FormAccessorBase<
     return this.formAccessor.isValid;
   }
 
-  access(name: string): Accessor | undefined {
+  @computed
+  get disabled(): boolean {
+    return this.formAccessor.disabled;
+  }
+
+  @computed
+  get hidden(): boolean {
+    return this.formAccessor.hidden;
+  }
+
+  @computed
+  get readOnly(): boolean {
+    return this.formAccessor.readOnly;
+  }
+
+  @computed
+  get inputAllowed(): boolean {
+    return this.formAccessor.inputAllowed;
+  }
+
+  @computed
+  get externalErrors(): ExternalMessages {
+    return this.formAccessor.externalErrors;
+  }
+
+  @computed
+  get externalWarnings(): ExternalMessages {
+    return this.formAccessor.externalWarnings;
+  }
+
+  @computed
+  get addMode(): boolean {
+    return this.formAccessor.addMode;
+  }
+
+  clear() {
+    return this.formAccessor.clear();
+  }
+
+  access(name: string): IAccessor | undefined {
     return this.formAccessor.access(name);
   }
 
-  accessBySteps(steps: string[]): Accessor | undefined {
+  accessBySteps(steps: string[]): IAccessor | undefined {
     if (steps.length === 0) {
       if (this.formAccessor.parent == null) {
         throw new Error("Unknown parent");
@@ -86,12 +126,12 @@ export abstract class FormAccessorBase<
   }
 
   @computed
-  get accessors(): Accessor[] {
+  get accessors(): IAccessor[] {
     return this.formAccessor.accessors;
   }
 
   @computed
-  get flatAccessors(): Accessor[] {
+  get flatAccessors(): IAccessor[] {
     return this.formAccessor.flatAccessors;
   }
 }