+++ src/repeating-form-accessor.ts
@@ -2,21 +2,24 @@ import { observable, computed } from "mobx";
 import { applyPatch } from "mobx-state-tree";
 import { FormDefinition, RepeatingForm, GroupDefinition } from "./form";
 import { FormState } from "./state";
-import { Accessor } from "./accessor";
 import { RepeatingFormIndexedAccessor } from "./repeating-form-indexed-accessor";
 import { FormAccessor } from "./form-accessor";
 import { ValidateOptions } from "./validate-options";
 import { pathToFieldref } from "./utils";
 import { ExternalMessages } from "./validationMessages";
+import { IAccessor } from "./interfaces";
 
 export class RepeatingFormAccessor<
   D extends FormDefinition<any>,
   G extends GroupDefinition<D>
-> {
+> implements IAccessor {
   name: string;
 
   @observable
-  repeatingFormIndexedAccessors: Map<number, any> = observable.map();
+  repeatingFormIndexedAccessors: Map<
+    number,
+    RepeatingFormIndexedAccessor<D, G>
+  > = observable.map();
 
   externalErrors = new ExternalMessages();
   externalWarnings = new ExternalMessages();
@@ -128,19 +131,14 @@ export class RepeatingFormAccessor<
 
   @computed
   get accessors(): RepeatingFormIndexedAccessor<D, G>[] {
-    // we get the entries in this map, in order of index
-    const length = Array.from(this.repeatingFormIndexedAccessors.values())
-      .length;
-    const result = [];
-    for (let i = 0; i < length; i++) {
-      result.push(this.repeatingFormIndexedAccessors.get(i));
-    }
+    const result = Array.from(this.repeatingFormIndexedAccessors.values());
+    result.sort((first, second) => first.index - second.index);
     return result;
   }
 
   @computed
-  get flatAccessors(): Accessor[] {
-    const result: Accessor[] = [];
+  get flatAccessors(): IAccessor[] {
+    const result: IAccessor[] = [];
     this.accessors.forEach(accessor => {
       result.push(...accessor.flatAccessors);
       result.push(accessor);
@@ -148,7 +146,7 @@ export class RepeatingFormAccessor<
     return result;
   }
 
-  accessBySteps(steps: string[]): Accessor | undefined {
+  accessBySteps(steps: string[]): IAccessor | undefined {
     const [first, ...rest] = steps;
     const nr = parseInt(first, 10);
     if (isNaN(nr)) {