+++ src/repeating-form-indexed-accessor.ts
@@ -7,12 +7,12 @@ import { setAddModeDefaults } from "./addMode";
 import { FormAccessorBase } from "./form-accessor-base";
 import { FormAccessor } from "./form-accessor";
 import { pathToFieldref } from "./utils";
-import { ExternalMessages } from "./validationMessages";
+import { IFormAccessor } from "./interfaces";
 
 export class RepeatingFormIndexedAccessor<
   D extends FormDefinition<any>,
   G extends GroupDefinition<D>
-> extends FormAccessorBase<D, G> {
+> extends FormAccessorBase<D, G> implements IFormAccessor<D, G> {
   formAccessor: FormAccessor<D, G>;
 
   @observable
@@ -20,9 +20,6 @@ export class RepeatingFormIndexedAccessor<
 
   @observable
   _addMode: boolean = false;
-
-  externalErrors = new ExternalMessages();
-  externalWarnings = new ExternalMessages();
 
   constructor(
     public state: FormState<any, any, any>,