+++ src/sub-form-accessor.ts
@@ -5,16 +5,13 @@ import { FormAccessor } from "./form-accessor";
 import { FormAccessorBase } from "./form-accessor-base";
 import { ValidateOptions } from "./validate-options";
 import { pathToFieldref } from "./utils";
-import { ExternalMessages } from "./validationMessages";
+import { IFormAccessor } from "./interfaces";
 
 export class SubFormAccessor<
   D extends FormDefinition<any>,
   G extends GroupDefinition<D>
-> extends FormAccessorBase<D, G> {
+> extends FormAccessorBase<D, G> implements IFormAccessor<D, G> {
   formAccessor: FormAccessor<D, G>;
-
-  externalErrors = new ExternalMessages();
-  externalWarnings = new ExternalMessages();
 
   constructor(
     public state: FormState<any, any, any>,